:wrench: ECS 패턴 적용, GUI 구조 분리
