+++ core/Cargo.toml
@@ -3,6 +3,38 @@ name = "core"
 version = "0.1.0"
 edition = "2021"
 
+[lib]
+crate-type = ['cdylib']
+
+[build]
+target = ["wasm32-unknown-unknown"]
+rustflags = ["-Zstrip=symbols"]
+
+[profile.release]
+lto = "thin"
+
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+wasm-bindgen = "0.2.81"
+specs = { version = "0.18.0", features = ["specs-derive"] }
+gui = { path = "../gui" }
+terrain = { path = "../terrain"}
+
+[dependencies.web-sys]
+version = "0.3.4"
+features = [
+  'Document',
+  'Element',
+  'CssStyleDeclaration',
+  'HtmlCanvasElement',
+  'HtmlInputElement',
+  'WebGlBuffer',
+  'WebGlVertexArrayObject',
+  'WebGl2RenderingContext',
+  'WebGlProgram',
+  'WebGlShader',
+  'WebGlUniformLocation',
+  'Window',
+  'console'
+]
