+++ core/src/components/terrain/mod.rs
@@ -1,0 +1,8 @@
+use specs::{VecStorage, Component};
+use terrain::model::pixel::Pixel;
+
+#[derive(Component, Debug)]
+#[storage(VecStorage)]
+pub struct Terrain {
+    pub bitmap: Vec<Vec<Pixel>>
+}
