+++ core/src/resources/physics/time/mod.rs
@@ -1,0 +1,5 @@
+#[derive(Default, Debug)]
+pub struct Time {
+    pub dt: u32,
+    pub time: std::time::Duration,
+}
