+++ gui/Cargo.toml
@@ -3,16 +3,6 @@ name = "gui"
 version = "0.1.0"
 edition = "2021"
 
-[lib]
-crate-type = ['cdylib']
-
-[build]
-target = ["wasm32-unknown-unknown"]
-rustflags = ["-Zstrip=symbols"]
-
-[profile.release]
-lto = "thin"
-
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
