+++ public/index.html
@@ -9,6 +9,10 @@
       margin: 0;
       padding: 0;
     }
+    canvas {
+      width: 100vw;
+      height: 100vh;
+    }
     #range_wrapper {
       position: fixed;
       display: block;
@@ -16,7 +20,7 @@
       right: 15px;
       padding: 20px;
       font-weight: 700;
-      background-color: rgba(0, 0, 0, 0.4);
+      background-color: rgba(255, 255, 255, 0.1);
       border-radius: 10px;
       color: #fff;
     }
